LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/transition-base-response-003-expected.txt


FAIL Identical elements are all responsive to font-size transition assert_equals: expected "15px" but got "20px"
